+++ common/src/main/java/com/evandev/modulation/mixin/vanilla/accessor/BlockDisplayInvoker.java
@@ -1,4 +1,4 @@
-package com.evandev.modulation.mixin.minecraft.accessor;
+package com.evandev.modulation.mixin.vanilla.accessor;
 
 import net.minecraft.world.entity.Display;
 import net.minecraft.world.level.block.state.BlockState;