+++ common/src/main/java/com/evandev/modulation/mixin/vanilla/client/GuiGraphicsMixin.java
@@ -1,4 +1,4 @@
-package com.evandev.modulation.mixin.minecraft.client;
+package com.evandev.modulation.mixin.vanilla.client;
 
 import com.evandev.modulation.Constants;
 import com.evandev.modulation.api.ModuleManager;