+++ common/src/main/java/com/evandev/modulation/mixin/vanilla/client/LocalPlayerMixin.java
@@ -1,4 +1,4 @@
-package com.evandev.modulation.mixin.minecraft.client;
+package com.evandev.modulation.mixin.vanilla.client;
 
 import com.evandev.modulation.registry.ModRegistry;
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;