+++ common/src/main/java/com/evandev/modulation/mixin/vanilla/perf/DensityFunctionsMapAllMixin.java
@@ -1,4 +1,4 @@
-package com.evandev.modulation.mixin.minecraft.perf;
+package com.evandev.modulation.mixin.vanilla.perf;
 
 import com.evandev.modulation.api.ModuleManager;
 import com.evandev.modulation.modules.VanillaModule;