+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2026-07-03
+
+### Added
+
+- Added an option for Campfires to place unlit.
+- Added an option for flammable cobwebs.
+
+### Changed
+
+- Switched from Cloth Config to YACL.
+- Backend rewrites.
+
 ## [1.6.2] - 2026-07-02
 
 ### Fixed